[NFC][RequirementMachine] Rename some of the RequirementError::Kind cases, add documentation for each case, and move RequirementError to lib/AST/RequirementMachine/Diagnostics.h.
